+++ prometheus/src/test/scala/fr/davit/akka/http/metrics/prometheus/marshalling/PrometheusMarshallersSpec.scala
@@ -20,11 +20,10 @@ import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import fr.davit.akka.http.metrics.core.HttpMetricsRegistry.StatusGroupDimension
 import fr.davit.akka.http.metrics.core.scaladsl.server.HttpMetricsDirectives.metrics
-import fr.davit.akka.http.metrics.core.scaladsl.server.HttpMetricsSettings
-import fr.davit.akka.http.metrics.prometheus.PrometheusRegistry
+import fr.davit.akka.http.metrics.prometheus.{PrometheusRegistry, PrometheusSettings}
 import io.prometheus.client.CollectorRegistry
-import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.BeforeAndAfterAll
+import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 
 import scala.concurrent.duration._
@@ -35,7 +34,7 @@ class PrometheusMarshallersSpec extends AnyFlatSpec with Matchers with Scalatest
 
     val registry = PrometheusRegistry(
       new CollectorRegistry(),
-      PrometheusRegistry.defaultSettings.withIncludeStatusDimension(true)
+      PrometheusSettings.default.withIncludeStatusDimension(true)
     )
 
     io.prometheus.client.Counter
@@ -72,15 +71,15 @@ class PrometheusMarshallersSpec extends AnyFlatSpec with Matchers with Scalatest
       metrics should contain theSameElementsAs Seq(
         "akka_http_requests_active",
         "akka_http_requests_total",
-        "akka_http_requests_size_bytes",
+        "akka_http_requests_size_bytes_bucket",
         "akka_http_requests_size_bytes_count",
         "akka_http_requests_size_bytes_sum",
         "akka_http_responses_total",
         "akka_http_responses_errors_total",
-        "akka_http_responses_duration_seconds",
+        "akka_http_responses_duration_seconds_bucket",
         "akka_http_responses_duration_seconds_count",
         "akka_http_responses_duration_seconds_sum",
-        "akka_http_responses_size_bytes",
+        "akka_http_responses_size_bytes_bucket",
         "akka_http_responses_size_bytes_count",
         "akka_http_responses_size_bytes_sum",
         "akka_http_connections_active",