+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Update libraries
 - Decorrelate routes from registry
 - Add namespace setting
+- Add histogram config for prometheus registry
 
 ## 0.6.0 (2019-08-25)